+++ app/(home)/(default-home)/page.tsx
@@ -1,16 +1,10 @@
 import dynamic from 'next/dynamic';
 import HomeTopBannerCard from './_components/HomeTopBannerCard';
-import HomeWidgetSection from './_components/HomeWidgetSection';
 import BucketListCard from './_components/BucketListCard';
 import DefaultBucketListCard from './_components/DefaultBucketListCard';
-import { CardSkeleton } from '@/components/ui/skeleton';
 import SwiperWrapper from '@/components/SwiperWrapper';
 const ChallengeCard = dynamic(() => import('./_components/ChallengeCard'), {
   ssr: false
-});
-const HomeAiBannerCard = dynamic(() => import('./_components/HomeAiBannerCard'), {
-  ssr: false,
-  loading: () => <CardSkeleton />
 });
 
 const HomePage = () => {
@@ -25,10 +19,6 @@ const HomePage = () => {
           <BucketListCard />
           <ChallengeCard />
         </SwiperWrapper>
-      </section>
-      <section className='px-20'>
-        <HomeWidgetSection />
-        <HomeAiBannerCard />
       </section>
     </>
   );
