+++ app/asset-management/layout.tsx
@@ -1,5 +1,11 @@
 import React from 'react';
 import { DefaultHeader } from '@/components/header';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'PORKO 자산관리 페이지',
+  description: '자산관리 페이지'
+};
 
 const AssetLayout = ({ children }: { children: React.ReactNode }) => {
   return (
